% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utilities.R
\name{extract_polygon}
\alias{extract_polygon}
\title{Summarise spatial polygon data around sites,}
\usage{
extract_polygon(my_points, location_column, my_buffer, my_shape, layers = NULL)
}
\arguments{
\item{my_points}{a spatial points data.frame that contains the locations
you wish to summarize landcover data at.}

\item{location_column}{the name of the column in \code{my_points} that
specifies the names of each location. Can either be the column name
or an integer that specifies it's column location.}

\item{my_buffer}{The radius of a buffer around each point form which}

\item{my_shape}{A \code{SpatialPolygonsDataFrame} object of a shape file you want to extract}

\item{layers}{A character vector of column names from my_shape
to summarise. Optional. If NULL, all numeric
columns are summarised.}
}
\description{
Given a set of locations and a spatial polygons layer,
 this function xtract the median value from spatial polygons that intersect a
buffer around a site.
}
